// apps/sync-service/src/openapi/order-item.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * SMILE
 * OpenAPI spec version: 1.0.0
 */
import axios from 'axios';
import type {
  AxiosRequestConfig,
  AxiosResponse
} from 'axios';

export type PutV2OrderOrderIdBodyOrderItemsItem = {
  id?: number;
  material_id?: number;
  qty?: number;
  ordered_qty?: number;
  reason_id?: number;
  other_reason?: null;
  pieces_per_unit?: number;
  temperature_sensitive?: number;
  recommend?: number;
  on_hand_stock?: number;
  mode?: number;
  max?: number;
  code_kfa_product_template?: string;
};

export type PutV2OrderOrderIdBody = {
  order_items: PutV2OrderOrderIdBodyOrderItemsItem[];
};

export type PutV2OrderOrderId200 = { [key: string]: unknown };

export const getSmile = () => {
/**
 * @summary Ubah Item
 */
const putV2OrderOrderId = <TData = AxiosResponse<PutV2OrderOrderId200>>(
    orderId: number,
    putV2OrderOrderIdBody: PutV2OrderOrderIdBody, options?: AxiosRequestConfig
 ): Promise<TData> => {
    return axios.put(
      `/v2/order/${orderId}`,
      putV2OrderOrderIdBody,options
    );
  }

return {putV2OrderOrderId}};
export type PutV2OrderOrderIdResult = AxiosResponse<PutV2OrderOrderId200>